Handle problematic answers from the arxiv servers
The arxiv API seems quite unstable, as we obtained a lot of
query responses with either less entries than asked for, or
no entries at all. In both cases, we end up with an
incomplete list of papers. The concrete behavior of the
arxiv servers varies significantly, and is not reproducible,
but larger requests (usually connected to dates further in
the past) are more prone to be erroneous.