Set number of requested results to fix value
We set max_results to 800 which is slightly below the
allowed value of 1000 recommended by the arxiv API because
we encountered less problems with 800. Also with such a
large value of max_results we have in total less requests,
which means a faster execution of the script overall.
Previously, max_results was dynamically determined and was
often much smaller.

In particular for --date=new and --date=recent this
immensely speeds up our script.