Remove check for cutoff date
We will not check if the cutoff date lies less than two
years in the past anymore, because this check is not needed
any longer. Instead, we only allow for a fixed amount of
errors to occur (100 at the moment), after which we exit the
script, so we avoid any endless loops (implemented already
in a previous commit #c8a0166).