Change arXiv API to OAI-PMH harvester API
Because of persisting problems with the present arXiv API we
make now use of a different arXiv API, the OAI-PMH API:
    https://info.arxiv.org/help/oa/index.html

This API allows for retrieving large amounts of records
(i.e. papers) and seems much more stable than the previously
used API. One main difference lies in handling dates: When
requesting records since a given date, the server will
provide all records that have been modified on the server
since that date. While we will always get all "new" records
with such a request, we will also possibly get some or many
additional "old" records that have neither been submitted,
nor updated (by resubmission) in that peroid, but simply
have been modified in the database of the server. However,
we can simply sort out these records.

With the new API we have to modifiy the convenction of our
categories. For the categories you can combine arXiv groups,
archives, and categories:
    physics                 group
    physics:astro-ph        archive
    physics:astro-ph:EP     category

While implementing the new API we could restore the old
behavior of "resubmissions", i.e., including resubmissions
in the listing when the flag is True (instead of ONLY
showing resubmissions).

PLEASE NOTE that you have to update your categories to match
the new convention, e.g.
    astro-ph.EP => physics:astro-ph:EP
and that you have to rename the config entry:
    only_resubmissions => resubmissions